+++ RockFramework/Commands/GetNextMessageCommand.cs
@@ -8,7 +8,7 @@ namespace Rock.Commands
 {
     public class GetNextMessageCommand : BaseCommand
     {
-        public GetNextMessageCommand(string appId, short key)
+        public GetNextMessageCommand(string appId, byte key)
             : base(CommandType.GetNextMessage, appId, key)
         {
